Wave tracker Standalone + Peak Data analysis in IR
Added a new data analyser for IR data (hard to retrieve data this way).

Created a "wave tracker" program that simply prepares the data for a specific case in october 12 2024, where one can apply any treatment (highpass and sqrt normalisation) to the data.

There's also a plot waverose function allowing to visualise any square at any frame and check all the coherence and phase information. It's standalone because it works on its own and does not require any external data to work.