Automated the Box pipeline
The box looper program automatically gathers the data online (from the BOX folder) and converts it to the chosen filtering before applying the entire routine, returning the spatially averaged cwt coeffs and images at each scale for the dominant angle of the wavelet real and absolute parts.

The update includes some corrections to filtering ( the updated version has no black stains around drastic gradients ) and the new cwts ( multiplied by 2/Scale)

There's also the wave_tracker_displayed, which allows to track for a given set of scares, the coherence enveloppe of a certain scale.