Introduced automatization
The program now automatically retrieves data, renames it and places it in a specified folder.

The data can be treated with unique methods tied to each instrument and this is all modular

the process of CWT and coherence is joined in a single program, but there's also a program for single instrument retrieval.

figure annotation has also been updated and includes multi-peak sum of wave coefficients.